Core: Fix String compile issue with memcpy.
std::memcpy using GCC will throw a warning saying the u64
is too large a data type. Casted to u32 to meet the
criteria.

+++ Source/Core/Containers/String.hpp
@@ -138,7 +138,7 @@ private:
     {
         Clear();
         m_Data.Resize(Length + 1);
-        std::memcpy(m_Data.Data(), Data, sizeof(T) * Length);
+        std::memcpy(m_Data.Data(), Data, sizeof(T) * (u32)Length);
         m_Data[Length] = 0;
         return *this;
     }
